Added view pager to detail art object view. Extracted GalleryImageAdapter to be able to reuse it for both detail art object view and image view.
