+++ app/src/main/java/org/imozerov/streetartview/ui/detail/ImageViewActivity.kt
@@ -1,34 +1,20 @@
 package org.imozerov.streetartview.ui.detail
 
 import android.app.Activity
-import android.content.Context
 import android.content.Intent
 import android.os.Bundle
-import android.support.v4.view.PagerAdapter
 import android.support.v4.view.ViewPager
 import android.support.v7.app.AppCompatActivity
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.LinearLayout
 import kotlinx.android.synthetic.main.activity_image_view.*
-import kotlinx.android.synthetic.main.art_object_gallery_item.view.*
 import org.imozerov.streetartview.R
 import org.imozerov.streetartview.StreetArtViewApp
 import org.imozerov.streetartview.storage.IDataSource
-import org.imozerov.streetartview.ui.extensions.loadImage
 import javax.inject.Inject
 
 class ImageViewActivity : AppCompatActivity() {
 
-    companion object {
-        val EXTRA_IMAGE_CHOSEN_IN_VIEWPAGER = "EXTRA_IMAGE_CHOSEN_IN_VIEWPAGER"
-    }
-
     @Inject
     lateinit var dataSource: IDataSource
-
-    private var pagerPosition: Int? = null
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -36,70 +22,26 @@ class ImageViewActivity : AppCompatActivity() {
 
         (application as StreetArtViewApp).appComponent.inject(this)
 
-
         val artObjectId = intent.getStringExtra(DetailArtObjectActivity.EXTRA_KEY_ART_OBJECT_DETAIL_ID)
         val artObjectUi = dataSource.getArtObject(artObjectId)
 
-        image_viewpager.adapter = GalleryPagerAdapter(this, artObjectUi.picsUrls)
+        image_viewpager.adapter = GalleryPagerAdapter(this, artObjectUi.picsUrls, closePager)
         val imageChosen = intent.getIntExtra(DetailArtObjectActivity.EXTRA_IMAGE_CHOSEN_IN_DETAILS, 0)
         image_viewpager.currentItem = imageChosen
-
-        image_viewpager.addOnPageChangeListener(GalleryPagerListener());
     }
 
     override fun onBackPressed() {
-        closePager()
+        closePager(image_viewpager.currentItem)
     }
 
-    fun closePager() {
+    val closePager = fun(position: Int) {
         val returnIntent = Intent()
-        returnIntent.putExtra(EXTRA_IMAGE_CHOSEN_IN_VIEWPAGER, pagerPosition)
+        returnIntent.putExtra(EXTRA_IMAGE_CHOSEN_IN_VIEWPAGER, position)
         setResult(Activity.RESULT_OK, returnIntent)
         finish()
     }
 
-    internal inner class GalleryPagerAdapter(val context: Context, val images: List<String>) : PagerAdapter() {
-        var inflater: LayoutInflater
-
-        init {
-            inflater = context.getSystemService(Context.LAYOUT_INFLATER_SERVICE) as LayoutInflater
-        }
-
-        override fun getCount(): Int {
-            return images.size
-        }
-
-        override fun isViewFromObject(view: View, anyObject: Any): Boolean {
-            return view === anyObject as LinearLayout
-        }
-
-        override fun instantiateItem(container: ViewGroup, position: Int): Any {
-            val itemView = inflater.inflate(R.layout.art_object_gallery_item, container, false)
-            itemView.image_preview.loadImage(images[position])
-            itemView.image_preview.setOnClickListener {
-                closePager()
-            }
-            container.addView(itemView)
-            return itemView
-        }
-
-        override fun destroyItem(container: ViewGroup, position: Int, anyObject: Any) {
-            container.removeView(anyObject as LinearLayout)
-        }
-    }
-
-    internal inner class GalleryPagerListener() : ViewPager.OnPageChangeListener {
-
-        override fun onPageScrolled(position: Int, positionOffset: Float, positionOffsetPixels: Int) {
-
-        }
-
-        override fun onPageSelected(position: Int) {
-            pagerPosition = position
-        }
-
-        override fun onPageScrollStateChanged(state: Int) {
-
-        }
+    companion object {
+        val EXTRA_IMAGE_CHOSEN_IN_VIEWPAGER = "EXTRA_IMAGE_CHOSEN_IN_VIEWPAGER"
     }
 }
